+++ src/main/java/com/Ecomercial/CTTT2018/models/service/ProductService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface ProductService {
 
-    Optional<Product>getProductById(Integer id);
+    Optional<Product>getProductById(Long id);
 
     Optional<Product>getProductByName(String name);
